+++ changes.md
@@ -15,6 +15,7 @@
 - Removed `ResponseFormat::formatOther` as non-formattable responses are returned as they are.
 - `Dingo\Api\Routing\Controller` is now `Dingo\Api\Routing\ControllerTrait`, you can now your own base controller and simply use the trait.
 - Custom transformation layers must now implement `Dingo\Api\Transformer\TransformerInterface`.
+- Authentication providers must now implement the `Dingo\Api\Auth\ProviderInterface`.
 
 ### v0.6.5